cambios finales para el final 16 de octubre

+++ Programacion_laboratorio_III_2018/js/cargarPaginaPrincipal.js
@@ -6,9 +6,7 @@ window.onload = function() {
     });
     funcionAjax.then(function(dato) {
         document.getElementById("usuario").innerHTML = "<p id='Usuario' style='color:white;'><span class='glyphicon glyphicon-user'>" + dato.Usuario + "</span></p>";
-    }, function(dato) {
-        console.log("No se pudo cargar el administrador!");
-    });
+    }, function(dato) {});
 
 }
 
@@ -17,32 +15,37 @@ function CerrarSesion() {
         title: 'Desea cerrar Sesión?',
         type: 'warning',
         showCancelButton: true,
+        showCloseButton: true,
         confirmButtonColor: '#3085d6',
         cancelButtonColor: '#d33',
         confirmButtonText: 'Si, quiero cerrar sesión!',
         cancelButtonClass: 'btn btn-danger',
         cancelButtonText: 'No, no deseo cerrar sesión!'
-    }).then(function() {
-        //FALTA AJAX CON SESION - REGISTRAR SU SALIDA
-        let id = localStorage.getItem("idEmpleado");
-        var funcionAjax = $.ajax({
-            method: 'POST',
-            url: '../vendor/Login/CerrarSesion',
-            data: { idEmpleado: id }
-        });
-        funcionAjax.then(function(dato) {
-            if (dato.status == 200) {
-                swal('Usted ha cerrado su sesión!').then(function() {
-                    localStorage.clear();
-                    window.location.replace("../enlaces/login.html");
-                }, function() {
-                    swal("OCURRIO ALGO INESPERADO!");
-                });
-            } else if (dato.status == 400) {
-                swal("Hubo un error al cerrar sesión del usuario!");
-            }
-        }, function(dato) {
-            console.log("ERROR en la API " + dato);
-        });
+    }).then(function(result) {
+        if (result.value) {
+
+
+            //FALTA AJAX CON SESION - REGISTRAR SU SALIDA
+            var funcionAjax = $.ajax({
+                method: 'POST',
+                url: '../vendor/Login/CerrarSesion',
+            });
+
+
+            funcionAjax.then(function(dato) {
+                if (dato.status == 200) {
+                    swal('Usted ha cerrado su sesión!').then(function() {
+                        localStorage.clear();
+                        window.location.replace("../enlaces/login.html");
+                    }, function() {
+                        swal("OCURRIO ALGO INESPERADO!");
+                    });
+                } else if (dato.status == 400) {
+                    swal("Hubo un error al cerrar sesión del usuario!");
+                }
+            }, function(dato) {
+                console.log("ERROR en la API " + dato);
+            });
+        }
     });
 }
